requirements.txt
django
gunicorn
whitenoise
psycopg2-binary
requests
urllib3
django-environ
django-cors-headers
psycopg2--2.9.3